crane: move qwen3_tts WAV loading out of crane-core, drop hound
crane-core model methods should work with raw sample slices, not
files; WAV I/O belongs in crane. Moves the WAV reader/resampler
(load_wav_f32) from qwen3_tts::model into the shared
crane/src/audio/tts.rs alongside encode_wav/save_wav, and changes
Model::generate_voice_clone to accept ref_samples: &[f32] instead of
a file path. Updates the Tts trait impl and both voice-clone examples
to load the reference WAV before calling in.

This removes hound, rubato, and audioadapter-buffers from crane-core
entirely. crane-core no longer touches WAV files at all.
