+++ crane/src/audio/mod.rs
@@ -9,5 +9,5 @@ mod tts_voxtral;
 
 #[cfg(feature = "onnx")]
 pub use asr::AsrClient;
-pub use tts::{AudioInfo, Tts, TtsStream, VoiceInfo, encode_wav, pcm_f32_to_i16, save_wav};
+pub use tts::{AudioInfo, Tts, TtsStream, VoiceInfo, encode_wav, load_wav_f32, pcm_f32_to_i16, save_wav};
 pub use tts_client::TtsClient;
