+++ example/src/tts_voice_clone.rs
@@ -59,6 +59,8 @@ fn main() -> anyhow::Result<()> {
     println!("Reference audio : {ref_audio}");
     println!("Reference text  : {ref_text}");
 
+    let ref_samples = crane::audio::load_wav_f32(ref_audio, model.speaker_encoder_sample_rate())?;
+
     // ── Sentences to clone ─────────────────────────────────────────
     let examples: &[(&str, &str, &str)] = &[
         (
@@ -85,7 +87,7 @@ fn main() -> anyhow::Result<()> {
             top_p: Some(1.0),
             repetition_penalty: 1.05,
         };
-        let (audio, sr) = model.generate_voice_clone(text, lang, ref_audio, ref_text, &opts)?;
+        let (audio, sr) = model.generate_voice_clone(text, lang, &ref_samples, ref_text, &opts)?;
         let saved_path = crane::audio::save_wav(&audio, &output_path, sr)?;
         let elapsed = start.elapsed();
         println!("  Saved {saved_path} in {elapsed:.1?}");
